+++ app/ante/utils_test.go
@@ -2,7 +2,6 @@ package ante_test
 
 import (
 	"fmt"
-	"math/big"
 	"testing"
 	"time"
 
@@ -37,7 +36,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	suite.app = app.Setup(checkTx)
 	suite.app.Codec().RegisterConcrete(&sdk.TestMsg{}, "test/TestMsg", nil)
 
-	suite.ctx = suite.app.BaseApp.NewContext(checkTx, abci.Header{Height: 1, ChainID: "3", Time: time.Now().UTC()})
+	suite.ctx = suite.app.BaseApp.NewContext(checkTx, abci.Header{Height: 1, ChainID: "ethermint-3", Time: time.Now().UTC()})
 	suite.app.EvmKeeper.SetParams(suite.ctx, evmtypes.DefaultParams())
 
 	suite.anteHandler = ante.NewAnteHandler(suite.app.AccountKeeper, suite.app.EvmKeeper, suite.app.SupplyKeeper)
@@ -91,9 +90,9 @@ func newTestSDKTx(
 }
 
 func newTestEthTx(ctx sdk.Context, msg evmtypes.MsgEthereumTx, priv tmcrypto.PrivKey) (sdk.Tx, error) {
-	chainID, ok := new(big.Int).SetString(ctx.ChainID(), 10)
-	if !ok {
-		return nil, fmt.Errorf("invalid chainID: %s", ctx.ChainID())
+	chainIDEpoch, err := ethermint.ParseChainID(ctx.ChainID())
+	if err != nil {
+		return nil, err
 	}
 
 	privkey, ok := priv.(crypto.PrivKeySecp256k1)
@@ -101,8 +100,7 @@ func newTestEthTx(ctx sdk.Context, msg evmtypes.MsgEthereumTx, priv tmcrypto.Pri
 		return nil, fmt.Errorf("invalid private key type: %T", priv)
 	}
 
-	err := msg.Sign(chainID, privkey.ToECDSA())
-	if err != nil {
+	if err := msg.Sign(chainIDEpoch, privkey.ToECDSA()); err != nil {
 		return nil, err
 	}