+++ scripts/run-solidity-tests.sh
@@ -2,27 +2,29 @@
 
 export GOPATH=~/go
 export PATH=$PATH:$GOPATH/bin
-go build -o ./build/ethermintd ./cmd/ethermintd 
+go build -o ./build/ethermintd ./cmd/ethermintd
 go build -o ./build/ethermintcli ./cmd/ethermintcli
 mkdir $GOPATH/bin
 cp ./build/ethermintd $GOPATH/bin
 cp ./build/ethermintcli $GOPATH/bin
 
+CHAINID="ethermint-1337"
+
 cd tests-solidity
 
 if command -v yarn &> /dev/null; then
-	yarn install
-else 
-	curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-	echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
-	sudo apt update && sudo apt install yarn
-	yarn install
+    yarn install
+else
+    curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
+    echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
+    sudo apt update && sudo apt install yarn
+    yarn install
 fi
 
 chmod +x ./init-test-node.sh
 ./init-test-node.sh > ethermintd.log &
 sleep 5
-ethermintcli rest-server --laddr "tcp://localhost:8545" --unlock-key localkey,user1,user2 --chain-id 1337 --trace --wsport 8546 > ethermintcli.log &
+ethermintcli rest-server --laddr "tcp://localhost:8545" --unlock-key localkey,user1,user2 --chain-id $CHAINID --trace --wsport 8546 > ethermintcli.log &
 
 cd suites/initializable
 yarn test-ethermint
@@ -30,7 +32,7 @@ yarn test-ethermint
 ok=$?
 
 if (( $? != 0 )); then
-	echo "initializable test failed: exit code $?"
+    echo "initializable test failed: exit code $?"
 fi
 
 killall ethermintcli
@@ -43,7 +45,7 @@ exit $ok
 
 ./../../init-test-node.sh > ethermintd.log &
 sleep 5
-ethermintcli rest-server --laddr "tcp://localhost:8545" --unlock-key localkey,user1,user2 --chain-id 1337 --trace --wsport 8546 > ethermintcli.log &
+ethermintcli rest-server --laddr "tcp://localhost:8545" --unlock-key localkey,user1,user2 --chain-id $CHAINID --trace --wsport 8546 > ethermintcli.log &
 
 cd ../initializable-buidler
 yarn test-ethermint
@@ -51,7 +53,7 @@ yarn test-ethermint
 ok=$(($? + $ok))
 
 if (( $? != 0 )); then
-	echo "initializable-buidler test failed: exit code $?"
+    echo "initializable-buidler test failed: exit code $?"
 fi
 
 killall ethermintcli