+++ scripts/start.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 ethermintd --home /ethermint/node$ID/ethermintd/ start > ethermintd.log &
 sleep 5
-ethermintcli rest-server --laddr "tcp://localhost:8545" --chain-id 7305661614933169792 --trace > ethermintcli.log &
+ethermintcli rest-server --laddr "tcp://localhost:8545" --chain-id "ethermint-7305661614933169792" --trace > ethermintcli.log &
 tail -f /dev/null